Element: Refactor element package to be React Native friendly
