+++ packages/element/src/components/index.js
@@ -1,0 +1,2 @@
+export { default as pure } from './pure';
+export { default as RawHTML } from './raw-html';
